Move handlers for task nav buttons into TaskNav
